Data Acquisition-Processing/requirements.txt
pandas
requests
gnews
google-generativeai
python-dotenv
seaborn
matplotlib
scikit-learn
openai
